Fix AJAX - changed all import lines - renamed file
